move testresources to a module where we could share them
